RF: Consolodated text stim timing demos
compare_text_timing.py compares timing of TextBox, TextBox2, and TextStim in one script; and replaces the three text*_vs_text*.py demos
